// src/codetables/generated/core-cc-enumerations-vehicle-manufacturer-model.d.ts
export declare const coreCcEnumerationsVehicleManufacturerModel: ({
    modelCode: string;
    manufacturerCode: string;
    manufacturerName: string;
    modelName: string;
    gapoEnabled: boolean;
    orderBy: number;
    apiName: string;
    product: string;
    erased: boolean;
    createdAt: string;
    vehicleTypeCode: string;
    ckpCode?: undefined;
} | {
    modelCode: string;
    manufacturerCode: string;
    manufacturerName: string;
    modelName: string;
    gapoEnabled: boolean;
    orderBy: number;
    ckpCode: number;
    apiName: string;
    product: string;
    erased: boolean;
    createdAt: string;
    vehicleTypeCode: string;
})[];